perf(test): speed up Vitest bootstrap
