+++ vitest.unit.config.ts
@@ -2,11 +2,9 @@ import { defineConfig } from "vitest/config";
 import baseConfig from "./vitest.config.ts";
 
 const baseTest = (baseConfig as { test?: { include?: string[]; exclude?: string[] } }).test ?? {};
-const include = baseTest.include ?? [
-  "src/**/*.test.ts",
-  "extensions/**/*.test.ts",
-  "test/format-error.test.ts",
-];
+const include = (
+  baseTest.include ?? ["src/**/*.test.ts", "extensions/**/*.test.ts", "test/format-error.test.ts"]
+).filter((pattern) => !pattern.includes("extensions/"));
 const exclude = baseTest.exclude ?? [];
 
 export default defineConfig({
